# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amazon Price Tracker API
  version: "1.0.0"
  description: |
    A WordPress REST API plugin for tracking Amazon product prices across multiple
    international marketplaces. Integrates with Amazon's Creators API to fetch product
    data and pricing, storing historical price records over time.

    ## Authentication

    All endpoints except `GET /health` require authentication. This API relies entirely on
    WordPress's own authentication layer - there is no bespoke API-key system. The two
    supported mechanisms are:

    - **Application Passwords (HTTP Basic Auth)** - the practical choice for external REST
      clients (Insomnia, Postman, curl, server-to-server integrations). Generate one under
      the target WordPress user's profile ("Application Passwords" section), then send it as
      `Authorization: Basic base64(username:application_password)`. Requires either HTTPS or
      `WP_ENVIRONMENT_TYPE` set to `local` (WordPress disables Application Passwords over
      plain HTTP otherwise, except for local environments).
    - **Cookie + nonce auth** - used implicitly by same-origin JavaScript running inside
      wp-admin (e.g. `wp.apiFetch`). Not practical for external clients and not modeled
      separately below; treat `basicAuth` as the representative scheme.

    Endpoints marked "Admin only" additionally require the authenticated user to have the
    `manage_options` WordPress capability (i.e. be an Administrator).

...

...

...
    ```

...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Get the current user's Amazon Creators API settings
      description: "`creators_credential_id` is returned masked (e.g. `amzn****mple`); `creators_credential_secret` is never returned."
      operationId: getSettings
      responses:
        "200":
          description: Current settings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
              example:
                id: 1
                user_id: 1
                creators_credential_id: "amzn****mple"
                creators_credential_version: "3.2"
                partner_tags: { UK: "getyourself0a-21" }
                created_at: "2026-07-13T17:38:12+00:00"
                updated_at: "2026-07-13T17:38:12+00:00"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No settings configured for this user yet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    put:
      tags: [Settings]
      summary: Create or update the current user's Amazon Creators API settings
      description: |
        On first setup, `creators_credential_id`, `creators_credential_secret` and
        `creators_credential_version` are all required. On subsequent calls, any
        omitted field is left unchanged; `partner_tags` is **merged** into the
        existing map (per-region keys are overwritten, others untouched) rather than
        replaced wholesale - use `DELETE /settings/partner-tags/{region}` to remove one.
      operationId: updateSettings
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSettingsRequest" }
            example:
              creators_credential_id: "amzn1.application-oa2-client.example"
              creators_credential_secret: "amzn1.oa2-cs.v1.example"
              creators_credential_version: "3.2"
              partner_tags: { UK: "test-21" }
      responses:
        "200":
          description: Settings updated (already existed).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        "201":
          description: Settings created (first-time setup).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        "400":
          description: A required credential field was missing on first setup, an invalid `creators_credential_version`, or an invalid entry in `partner_tags`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /settings/validate:
    post:
      tags: [Settings]
      summary: Validate the current user's Amazon Creators API credentials
      description: Thin wrapper around the same connectivity test as `GET /health/amazon`.
      operationId: validateSettings
      responses:
        "200":
          description: Validation result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  message: { type: string }
              examples:
                valid:
                  value: { valid: true, message: Successfully connected to Amazon Creators API }
                invalid:
                  value: { valid: false, message: "The request signature we calculated does not match the signature you provided." }
        "400":
          description: No credentials configured at all (`NOT_CONFIGURED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ============================================================
  # Stats
  # ============================================================
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        WordPress Application Password. Username is the WordPress user's login name;
        password is the generated Application Password (spaces optional).
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example:
            code: rest_forbidden
            message: Authentication required.
            data: { status: 401 }
...
  schemas:
...
    Settings:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        creators_credential_id:
          type: string
          description: Creators API credential ID, masked (first/last 4 characters visible). Empty string if not configured. `creators_credential_secret` is never returned by any endpoint.
        creators_credential_version:
          type: string
          description: Creators API credential version (e.g. `"3.2"`), returned as plain text - not a secret, identifies which auth flavor/regional cluster the credential belongs to. Empty string if not configured.
        partner_tags:
          type: object
          description: Map of region code to Amazon Associates partner/tracking tag.
          additionalProperties: { type: string }
          example: { UK: "getyourself0a-21", US: "myaccount-20" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UpdateSettingsRequest:
      type: object
      properties:
        creators_credential_id:
          type: string
          description: Creators API credential ID (from Associates Central, Tools > Creators API).
        creators_credential_secret:
          type: string
          description: Creators API credential secret. Write-only - never echoed back in responses.
        creators_credential_version:
          type: string
          description: One of `2.1`, `2.2`, `2.3` (Cognito) or `3.1`, `3.2`, `3.3` (Login-with-Amazon). An optional leading `v` is accepted and stripped.
        partner_tags:
          type: object
          additionalProperties: { type: string }
          description: Merged into any existing partner tags map; per-region keys are overwritten.
...
    ValidationErrorDetail:
      type: object
      properties:
        field: { type: string }
        message: { type: string }
    ErrorResponse:
      type: object
      description: Standard WordPress `WP_Error` REST serialization.
      required: [code, message, data]
      properties:
        code:
          type: string
          description: Machine-readable error code (e.g. `VALIDATION_ERROR`, `ASIN_NOT_FOUND`, `RATE_LIMIT_EXCEEDED`, `NOT_FOUND`, `FORBIDDEN`, `rest_forbidden`).
        message:
          type: string
        data:
          type: object
          required: [status]
          properties:
            status: { type: integer, description: Mirrors the HTTP status code. }
            errors:
              type: array
              items: { $ref: "#/components/schemas/ValidationErrorDetail" }
              description: Present only for `VALIDATION_ERROR`.
            details:
              type: object
              description: Present only for a few error codes (e.g. `ALREADY_EXISTS` includes `{id}`, `BLACKLISTED` includes `{reason}`).
            limit: { type: integer, description: Present only for `RATE_LIMIT_EXCEEDED`. }
            used: { type: integer, description: Present only for `RATE_LIMIT_EXCEEDED`. }
            resets_at: { type: string, format: date-time, description: Present only for `RATE_LIMIT_EXCEEDED`. }